[#67] 데이터 없을때도 캘린더 보이게 수정
